+++ backend/pydantic_schemas/update_schemas/kidney_update_schema.py
@@ -5,9 +5,9 @@ import datetime
 
 class Kidney_Update(BaseModel):
     k_bu : Optional[int]
-    k_cc : Optional[int]
+    k_cc : Optional[float]
     k_ss : Optional[int]
-    k_sp : Optional[int]
+    k_sp : Optional[float]
     k_sc : Optional[int]
-    k_sca : Optional[int]
-    k_ur : Optional[int]
+    k_sca : Optional[float]
+    k_ur : Optional[float]
