+++ backend/pydantic_schemas/update_schemas/liver_update_schema.py
@@ -4,9 +4,9 @@ from pydantic import datetime_parse
 import datetime
 
 class Liver_Update(BaseModel):
-    itb : Optional[int]
-    idb : Optional[int]
-    iib : Optional[int]
+    itb : Optional[float]
+    idb : Optional[float]
+    iib : Optional[float]
     isgot : Optional[int]
     isgpt : Optional[int]
     iap : Optional[int]
